chore: move record type package
