+++ app/src/main/java/com/gdsc/addi/ui/feature/record/Record.kt
@@ -16,7 +16,6 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gdsc.addi.R
-import com.gdsc.addi.ui.RecordType
 import com.gdsc.addi.ui.component.AddiButton
 import com.gdsc.addi.ui.component.RecordButton
 import com.gdsc.addi.ui.theme.AddiDesignSystem
