+++ app/src/main/java/com/gdsc/addi/ui/feature/record/RecordDone.kt
@@ -14,7 +14,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.gdsc.addi.ui.RecordType
 import com.gdsc.addi.ui.component.AddiButton
 import com.gdsc.addi.ui.theme.AddiDesignSystem
 import com.gdsc.addi.ui.theme.AddiTheme
